WIP: Implement Andor camera control via prompt API
Andor cameras require a proprietary driver that is only available on x86
Windows and Linux and therefore can't be used directly on the Pi.
Instead, the GUI exposes access to the camera over the prompt API. The
HAL continues to be primarily responsible for running image sequences
and manages the GPIO.

This change implements the initial setup, image capture, and teardown of
the camera. The following work is NYI:

* Saving the images
* Image preview (particularly important because this deprecates usage of
  SOLIS alongside the GUI)